Bug #1391094 soc: sprd: debug: Set intervel to unsigned int
[root cause  ] implicit conversion (call) from 'u64' to 'unsigned int'
[changes     ] set intervel to unsigned int
[side effects] n/a
[self test   ]: power on
[download normally]: yes
[power on/off normally]: yes
[do common repository/branch inspection]: yes
[is there dependence]: no
[confirm dependent commit]: n/a
[board]: 7731e
[test case]: hang debug
[reviewers   ] orson.zhai
[change_type ] feature
[tag_product ] common

Intervel implicit conversion (call) from 'u64' to 'unsigned int',
so,set intervel to unsigned int to forced conversion.

+++ drivers/soc/sprd/debug/irq/irqs_monitor/irq_monitor.c
@@ -413,7 +413,7 @@ static ssize_t time_interval_write(struct file *filep,
 		threshold_irq = (unsigned int)mid_value;
 		spin_unlock(&irq_monitor_lock);
 
-		irq_monitor_timer->expires = jiffies + msecs_to_jiffies(interval);
+		irq_monitor_timer->expires = jiffies + msecs_to_jiffies((unsigned int)interval);
 		add_timer(irq_monitor_timer);
 	}
